requirements.txt
asgiref==3.8.1
Django==4.2.11
django-filter==24.2
django-shell-plus==1.1.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
